stress less haxe server + cache haxelib + usage of tmp dir
- optimize call to stress less haxe server, and try to cancel early some
diagnostic if the file is still editing
- a new interal build file is created (default name :
"vscode-project.hxml", configurable with the option
haxe.haxeVSCodeBuildFile) this build file is actually used to cache
haxelib call, but in the future will serve for building / selecting
target...
- if the option haxe.haxeCacheHaxelib is set to true haxelib call will
be cache into the internal build file, it is recached at restart and
each time your build file changed
- without modifying haxe server you can use now, with the option
haxe.haxeTmpDirectory, a temporary directory where the file you are
editing are saved, the original one remained untouched unless you save
it by yourself. For speeding thing you can setup a ramdisk and set the
option to this disk (on windows 7 i setup one with imdisk)
- with the option haxe.haxeUseTmpAsWorkingDirectory you can tell if you
want also the tmp directory to be your working directory.